chore: more RFCCommon refactoring, especially subseries
